+++ projects/stm32f407g-disc1/blinky/src/stm_interface.h
@@ -1,0 +1,6 @@
+#define STM32F407xx
+#define USE_HAL_DRIVER
+
+
+#include "main.h"
+#include "usb_host.h"