+++ projects/stm32l476_nucleo/blinky/build.zig
@@ -17,20 +17,33 @@ pub fn build(b: *std.Build) void {
 
     // Standard release options allow the person running `zig build` to select
     // between Debug, ReleaseSafe, ReleaseFast, and ReleaseSmall.
-    const optimization = b.standardOptimizeOption(.{});
+    const optimize = b.standardOptimizeOption(.{});
 
     // In Debug Release, the default optimization level is set to -O0, which significantly increases the binary size.
     // We override the optimization level with -Og while keeping the other three optimization modes unchanged.
-    const c_optimization = if (optimization == .Debug) "-Og" else if (optimization == .ReleaseSmall) "-Os" else "-O2";
+    const c_optimize = if (optimize == .Debug) "-Og" else if (optimize == .ReleaseSmall) "-Os" else "-O2";
+
+    const translate_c = b.addTranslateC(.{
+        .root_source_file = b.path("src/stm_interface.h"),
+        .target = target,
+        .optimize = optimize,
+        .link_libc = false,
+    });
 
     const exe_mod = b.createModule(.{
         .root_source_file = b.path("src/main.zig"),
         .target = target,
-        .optimize = optimization,
+        .optimize = optimize,
         .link_libc = false,
         .strip = false,
         .single_threaded = true, // single core cpu
         .sanitize_c = .trap,
+        .imports = &.{
+            .{
+                .name = "c",
+                .module = translate_c.createModule(),
+            },
+        },
     });
 
     const elf = b.addExecutable(.{
@@ -90,7 +103,7 @@ pub fn build(b: *std.Build) void {
     ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
     const hal_mod = b.createModule(.{
         .target = target,
-        .optimize = optimization,
+        .optimize = optimize,
         .link_libc = false,
         .strip = false,
         .single_threaded = true, // single core cpu
@@ -107,6 +120,7 @@ pub fn build(b: *std.Build) void {
 
     for (hal_includes) |path| {
         hal_mod.addIncludePath(b.path(path));
+        translate_c.addIncludePath(b.path(path));
     }
 
     hal_mod.addCSourceFiles(.{
@@ -132,7 +146,7 @@ pub fn build(b: *std.Build) void {
             "Drivers/STM32L4xx_HAL_Driver/Src/stm32l4xx_hal_exti.c",
         },
         .flags = &.{
-            c_optimization,
+            c_optimize,
             "-std=gnu17",
             "-Wall",
             "-Wextra",
@@ -157,7 +171,7 @@ pub fn build(b: *std.Build) void {
             "Core/Src/syscalls.c",
         },
         .flags = &.{
-            c_optimization,
+            c_optimize,
             "-std=gnu17",
             "-Wall",
             "-Wextra",
@@ -173,6 +187,7 @@ pub fn build(b: *std.Build) void {
     };
     for (app_includes) |path| {
         exe_mod.addIncludePath(b.path(path));
+        translate_c.addIncludePath(b.path(path));
     }
 
     exe_mod.addAssemblyFile(b.path("startup_stm32l476xx.s"));