+++ projects/stm32l476_nucleo/blinky/src/main.zig
@@ -1,10 +1,5 @@
 const std = @import("std");
-const c = @cImport({
-    @cDefine("USE_HAL_DRIVER", {});
-    @cDefine("STM32L476xx", {});
-    @cInclude("main.h");
-    @cInclude("usart.h");
-});
+const c = @import("c");
 
 export fn zigEntrypoint() callconv(.c) noreturn {
     while (true) {