+++ projects/stm32l476_nucleo/blinky/src/stm_interface.h
@@ -1,0 +1,5 @@
+#define STM32L476xx
+#define USE_HAL_DRIVER
+
+#include "main.h"
+#include "usart.h"