+++ projects/stm32l476_nucleo/blinky_freertos/build.zig
@@ -17,20 +17,43 @@ pub fn build(b: *std.Build) void {
 
     // Standard release options allow the person running `zig build` to select
     // between Debug, ReleaseSafe, ReleaseFast, and ReleaseSmall.
-    const optimization = b.standardOptimizeOption(.{});
+    const optimize = b.standardOptimizeOption(.{});
 
     // In Debug Release, the default optimization level is set to -O0, which significantly increases the binary size.
     // We override the optimization level with -Og while keeping the other three optimization modes unchanged.
-    const c_optimization = if (optimization == .Debug) "-Og" else if (optimization == .ReleaseSmall) "-Os" else "-O2";
+    const c_optimize = if (optimize == .Debug) "-Og" else if (optimize == .ReleaseSmall) "-Os" else "-O2";
+
+    const translate_c = b.addTranslateC(.{
+        .root_source_file = b.path("src/stm_interface.h"),
+        .target = target,
+        .optimize = optimize,
+        .link_libc = false,
+    });
+    const translate_os = b.addTranslateC(.{
+        .root_source_file = b.path("src/stm_os.h"),
+        .target = target,
+        .optimize = optimize,
+        .link_libc = false,
+    });
 
     const exe_mod = b.createModule(.{
         .root_source_file = b.path("src/main.zig"),
         .target = target,
-        .optimize = optimization,
+        .optimize = optimize,
         .link_libc = false,
         .strip = false,
         .single_threaded = true, // single core cpu
         .sanitize_c = .trap,
+        .imports = &.{
+            .{
+                .name = "c",
+                .module = translate_c.createModule(),
+            },
+            .{
+                .name = "os",
+                .module = translate_os.createModule(),
+            },
+        },
     });
 
     const elf = b.addExecutable(.{
@@ -91,7 +114,7 @@ pub fn build(b: *std.Build) void {
     ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
     const hal_mod = b.createModule(.{
         .target = target,
-        .optimize = optimization,
+        .optimize = optimize,
         .link_libc = false,
         .strip = false,
         .single_threaded = true, // single core cpu
@@ -128,7 +151,7 @@ pub fn build(b: *std.Build) void {
         "Drivers/STM32L4xx_HAL_Driver/Src/stm32l4xx_hal_exti.c",
     };
     const hal_flags = [_][]const u8{
-        c_optimization,
+        c_optimize,
         "-std=gnu17",
         "-Wall",
         "-Wextra",
@@ -136,6 +159,7 @@ pub fn build(b: *std.Build) void {
 
     for (hal_includes) |path| {
         hal_mod.addIncludePath(b.path(path));
+        translate_c.addIncludePath(b.path(path));
     }
 
     hal_mod.addCSourceFiles(.{
@@ -152,7 +176,7 @@ pub fn build(b: *std.Build) void {
     // FreeRTOS source code
     const os_mod = b.createModule(.{
         .target = target,
-        .optimize = optimization,
+        .optimize = optimize,
         .link_libc = false,
         .strip = false,
         .single_threaded = true, // single core cpu
@@ -183,7 +207,7 @@ pub fn build(b: *std.Build) void {
         "Middlewares/Third_Party/FreeRTOS/Source/portable/GCC/ARM_CM4F/port.c",
     };
     const os_flags = [_][]const u8{
-        c_optimization,
+        c_optimize,
         "-std=gnu17",
         "-Wall",
         "-Wextra",
@@ -191,6 +215,7 @@ pub fn build(b: *std.Build) void {
 
     for (os_includes) |path| {
         os_mod.addIncludePath(b.path(path));
+        translate_os.addIncludePath(b.path(path));
     }
     os_mod.addCSourceFiles(.{
         .files = &os_sources,
@@ -218,6 +243,7 @@ pub fn build(b: *std.Build) void {
     };
     for (app_includes) |path| {
         exe_mod.addIncludePath(b.path(path));
+        translate_c.addIncludePath(b.path(path));
     }
 
     const app_sources = [_][]const u8{
@@ -234,7 +260,7 @@ pub fn build(b: *std.Build) void {
         "Core/Src/freertos-openocd.c",
     };
     const app_flags = [_][]const u8{
-        c_optimization,
+        c_optimize,
         "-std=gnu17",
         "-Wall",
         "-Wextra",
@@ -247,6 +273,7 @@ pub fn build(b: *std.Build) void {
     const c_includes_core = [_][]const u8{"Core/Inc"};
     for (c_includes_core) |path| {
         exe_mod.addIncludePath(b.path(path));
+        translate_c.addIncludePath(b.path(path));
     }
 
     exe_mod.addAssemblyFile(b.path("startup_stm32l476xx.s"));