+++ projects/stm32l476_nucleo/blinky_freertos/src/stm_interface.h
@@ -1,0 +1,6 @@
+#define STM32L476xx
+#define USE_HAL_DRIVER
+
+
+#include "main.h"
+#include "usart.h"