+++ projects/stm32l476_nucleo/blinky_picolibc/src/stm_interface.h
@@ -1,0 +1,8 @@
+#define STM32L476xx
+#define USE_HAL_DRIVER
+
+#define _PICOLIBC_PRINTF m
+#define _PICOLIBC_SCANF m
+
+#include "main.h"
+#include "usart.h"